requirements.txt
djangorestframework==3.14.0
psycopg2-binary
python-dotenv
Pillow==9.3.0

